adds evm_client implementation
