+++ x/qgb/orchestrator/evm_transaction_opts.go
@@ -1,0 +1,92 @@
+package orchestrator
+
+import (
+	"context"
+	"crypto/ecdsa"
+	"errors"
+	"fmt"
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/accounts"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
+	ethcmn "github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+	ethcrypto "github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/ethclient"
+)
+
+// TODO: make gas price configurable
+type transactOpsBuilder func(ctx context.Context, client *ethclient.Client, gasLim uint64) (*bind.TransactOpts, error)
+
+func newTransactOptsBuilder(privKey *ecdsa.PrivateKey) transactOpsBuilder {
+	publicKey := privKey.Public()
+	publicKeyECDSA, ok := publicKey.(*ecdsa.PublicKey)
+	if !ok {
+		panic(fmt.Errorf("invalid public key; expected: %T, got: %T", &ecdsa.PublicKey{}, publicKey))
+	}
+
+	evmAddress := ethcrypto.PubkeyToAddress(*publicKeyECDSA)
+	return func(ctx context.Context, client *ethclient.Client, gasLim uint64) (*bind.TransactOpts, error) {
+		nonce, err := client.PendingNonceAt(ctx, evmAddress)
+		if err != nil {
+			return nil, err
+		}
+
+		ethChainID, err := client.ChainID(ctx)
+		if err != nil {
+			return nil, fmt.Errorf("failed to get Ethereum chain ID: %w", err)
+		}
+
+		auth, err := bind.NewKeyedTransactorWithChainID(privKey, ethChainID)
+		if err != nil {
+			return nil, fmt.Errorf("failed to create Ethereum transactor: %w", err)
+		}
+
+		bigGasPrice, err := client.SuggestGasPrice(ctx)
+		if err != nil {
+			return nil, fmt.Errorf("failed to get Ethereum gas estimate: %w", err)
+		}
+
+		auth.Nonce = new(big.Int).SetUint64(nonce)
+		auth.Value = big.NewInt(0) // in wei
+		auth.GasLimit = gasLim     // in units
+		auth.GasPrice = bigGasPrice
+
+		return auth, nil
+	}
+}
+
+type PersonalSignFn func(account ethcmn.Address, data []byte) (sig []byte, err error)
+
+func PrivateKeyPersonalSignFn(privKey *ecdsa.PrivateKey) (PersonalSignFn, error) {
+	keyAddress := crypto.PubkeyToAddress(privKey.PublicKey)
+
+	signFn := func(from common.Address, data []byte) (sig []byte, err error) {
+		if from != keyAddress {
+			return nil, errors.New("from address mismatch")
+		}
+
+		protectedHash := accounts.TextHash(data)
+		return crypto.Sign(protectedHash, privKey)
+	}
+
+	return signFn, nil
+}
+
+// SigToVRS breaks apart a signature into its components to make it compatible with the contracts
+func SigToVRS(sigHex string) (v uint8, r, s ethcmn.Hash) {
+	signatureBytes := ethcmn.FromHex(sigHex)
+	vParam := signatureBytes[64]
+	if vParam == byte(0) {
+		vParam = byte(27)
+	} else if vParam == byte(1) {
+		vParam = byte(28)
+	}
+
+	v = vParam
+	r = ethcmn.BytesToHash(signatureBytes[0:32])
+	s = ethcmn.BytesToHash(signatureBytes[32:64])
+
+	return
+}
